+++ tiling-assistant@leleat-on-github/prefs.js
@@ -166,7 +166,8 @@ export default class Prefs extends ExtensionPreferences {
                     'dynamic_keybinding_window_focus_row',
                     'dynamic_keybinding_tiling_state_row',
                     'dynamic_keybinding_tiling_state_windows_row',
-                    'dynamic_keybinding_favorite_layout_row'
+                    'dynamic_keybinding_favorite_layout_row',
+                    'dynamic_keybinding_edge_cycle_row'
                 ]
             },
             {